% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{gauss_legendre_01}
\alias{gauss_legendre_01}
\title{Gauss-Legendre nodes and weights on [0, 1]}
\usage{
gauss_legendre_01(M)
}
\arguments{
\item{M}{Number of nodes.}
}
\value{
A list of \code{gl_nodes_01} and \code{gl_weights_01}.
}
\description{
By the Golub-Welsch algorithm, mapped from [-1, 1] so the weights sum to
one and \code{sum(w * f(x))} approximates the integral of \code{f} over [0, 1].
}
\keyword{internal}
